Move news logic from frontend to backend and introduce NewsApi
