+++ server/App.ts
@@ -1,10 +1,9 @@
 import * as express from 'express';
 import splusController from './controllers/v1/splus';
 import mensaController from './controllers/mensa';
-import newsController from './controllers/v1/news';
+import newsController from './controllers/news';
 import icsController from './controllers/v1/ics';
 import splusController2 from './controllers/v2/splus';
-import newsController2 from './controllers/v2/news';
 import icsController2 from './controllers/v2/ics';
 
 class App {
@@ -21,7 +20,6 @@ class App {
     this.app.use(this.basePath + '/news', newsController);
     this.app.use(this.basePath + '/ics', icsController);
     this.app.use(this.basePath + '/v2/splus', splusController2);
-    this.app.use(this.basePath + '/v2/news', newsController2);
     this.app.use(this.basePath + '/v2/ics', icsController2);
   }
 }
