+++ store/news.js
@@ -1,5 +1,3 @@
-import { flatten } from "../lib/util";
-
 export const state = () => ({
   faculty: 'wf',
   campusNews: [],
@@ -20,16 +18,12 @@ export const mutations = {
 
 export const actions = {
   async loadFacultyNews({ state, commit }) {
-    const faculties = ['i', 'e', 'r', 'wf', 'wob', 'sz', 'sud'];
+    const faculties = ['i', 'r', 'e', 'sz', 'wf', 'wob', 'sud'];
     try {
-      const news = await Promise.all(
-        faculties.map(async (faculty) => ({
-          faculty,
-          articles: (await this.$axios.$get('/api/news/ostfalia/' + faculty)),
-        }))
-      );
+      const news = await this.$axios.$get('/api/news/faculty');
+
       const newsMap = {};
-      news.forEach(({ faculty, articles }) => newsMap[faculty] = articles);
+      faculties.forEach((faculty) => newsMap[faculty] = news.filter((article) => article.source == faculty));
 
       commit('setFacultyNews', newsMap);
     } catch (error) {
@@ -39,34 +33,9 @@ export const actions = {
   },
   async loadCampusNews({ state, commit }) {
     try {
-      const ostfaliaNews = await this.$axios.$get('/api/news/ostfalia');
-      const campus38News = await this.$axios.$get('/api/news/campus38');
+      const campusNews = await this.$axios.$get('/api/news/campus');
 
-      const truncateArticle = (article) => {
-        const sentences = article.text.split('.');
-        const text = sentences.reduce((text, sentence) => text.length < 50 ? text + sentence + '.' : text, '');
-        return {
-          ...article,
-          text,
-        };
-      };
-
-      const articles = [].concat(
-        ostfaliaNews.map((article) => ({ ...article, source: 'Ostfalia' })),
-        campus38News.map((article) => ({ ...article, source: 'Campus 38' }))
-      ).map(truncateArticle);
-
-      const scoreArticle = (article) => {
-        const date = new Date(article.date).getTime();
-        const now = new Date().getTime();
-        const age = now - date;
-        const boost = article.source == 'Ostfalia' ? 3.0 : 1.0;
-        return age / boost;
-      };
-
-      articles.sort((a1, a2) => scoreArticle(a1) - scoreArticle(a2));
-
-      commit('setCampusNews', articles);
+      commit('setCampusNews', campusNews);
     } catch (error) {
       commit('enqueueError', `News: API-Verbindung fehlgeschlagen (Ostfalia-News)`, {root: true});
       console.error(`error during News API call (Ostfalia-News)`, error.message);
